requirements.txt
jupyter>=1.0.0
jupyterlab>=4.0.0
notebook>=7.0.0

# LangSmith and LangChain
langsmith>=0.1.0
langchain>=0.1.0
langchain-openai>=0.1.0

# Data science and analysis
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Environment management
python-dotenv>=1.0.0

# HTTP requests
requests>=2.31.0

# OpenAI (for model evaluation)
openai>=1.0.0

# Vertex AI
langchain-google-vertexai>=0.0.1
google-cloud-aiplatform>=1.38.0
